Avances pre-exposi

+++ Mascoshop/Mascoshop/Mascoshop/src/main/java/com/Mascoshop/Entidades/Producto.java
@@ -17,7 +17,7 @@ public class Producto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_producto")
-    private long idProducto;
+    private Integer idProducto;
 
     @ManyToOne
     @JoinColumn(name = "idpcategoria")
